whole lot of changes
- fully implement saving requests
- change request form to be able to edit saved requests, create
new requests
- add diaglogue to insert/update request.

src/components/domain/request.history/request.history.filter.tsx
src/components/domain/request.history/request.history.item.actions.tsx
src/components/domain/request.saved/request.saved.item.actions.tsx
src/components/domain/request.section/request.actions.dropdown.tsx

+++ src/redux/request.history/history.async.actions.ts
@@ -1,14 +1,8 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
-import { RootState } from "../store";
 import { RequestModel } from "../../types";
+import { historyDb } from "../../lib/db";
 
-export const addtoHistoryAsync = createAsyncThunk<RequestModel, RequestModel>('history/addtoHistoryAsync', async (model, thunkAPI) => {
-    const { getState } = thunkAPI;
-
-    const rootState = getState() as RootState;
-
-    const historyDb = rootState.storage.historydb;
-
+export const addtoHistoryAsync = createAsyncThunk<RequestModel, RequestModel>('history/addtoHistoryAsync', async (model, _) => {
     //try to store request history item in indexedb
     try {
         const id = await historyDb?.insert(model);
@@ -20,54 +14,42 @@ export const addtoHistoryAsync = createAsyncThunk<RequestModel, RequestModel>('h
     return model;
 });
 
-export const loadHistoryFromDbAsync = createAsyncThunk<RequestModel[], void>("history/loadHistoryFromDbAsync", async (_, thunkAPI) => {
-
-    const { getState } = thunkAPI;
-
-    const rootState = getState() as RootState;
-
-    const historyDb = rootState.storage.historydb;
+export const loadHistoryFromDbAsync = createAsyncThunk<RequestModel[], void>("history/loadHistoryFromDbAsync", async (_, __) => {
 
     try {
-        const models = await historyDb?.getAll();
+        let models: RequestModel[] = [];
+        if (historyDb.isInitialized) {
+            models = await historyDb?.getAll();
+        }
 
-        return models || [];
+        return models;
 
     } catch (error) {
         console.log(`unable to load history items from db: ${error}`);
         return [];
     }
-
 });
 
-export const clearHistoryAsync = createAsyncThunk<void, void>("history/clearHistoryAsync", async (_, thunkAPI) => {
-
-    const { getState } = thunkAPI;
-
-    const rootState = getState() as RootState;
-
-    const historyDb = rootState.storage.historydb;
+export const clearHistoryAsync = createAsyncThunk<void, void>("history/clearHistoryAsync", async (_, __) => {
 
     try {
-        await historyDb?.deleteAll();
+        if (historyDb.isInitialized) {
+            await historyDb?.deleteAll();
+        }
     } catch (error) {
         console.log(`couldn't clear history of indexeddb`);
     }
 });
 
-export const deleteHistoryItemAsync = createAsyncThunk<string, string>("history/deleteHistoryItemAsync", async (id, thunkAPI) => {
-    const { getState } = thunkAPI;
-
-    const rootState = getState() as RootState;
-
-    const historyDb = rootState.storage.historydb;
+export const deleteHistoryItemAsync = createAsyncThunk<string, string>("history/deleteHistoryItemAsync", async (id, _) => {
 
     try {
-        await historyDb?.deleteById(id);
+        if (historyDb.isInitialized) {
+            await historyDb?.deleteById(id);
+        }
     } catch (error) {
         console.log(`couldn't delete item from history of indexeddb`);
     }
 
     return id;
-
 });
